Restructured directory organization, search bar styling
