+++ frontend/src/components/Pin/index.jsx
