--- frontend/src/styles/mapView.scss
@@ -1,6 +1,0 @@
-.search {
-    width: 20%;
-    position: absolute;
-    margin-left: 5%;
-    margin-top: 3%;
-}
