--- frontend/src/styles/search.scss
@@ -1,23 +1,0 @@
-.searchBar {
-  width: 100%;
-}
-
-.searchInput {
-  width: 100%;
-  height: 100%;
-  box-sizing: border-box;
-  border: 2px solid #ccc;
-  border-radius: 4px;
-  font-size: 16px;
-  background-color: white;
-  background-image: url('../images/searchicon.png');
-  background-position: 10px 9px; 
-  background-size: 15px 15px;
-  background-repeat: no-repeat;
-  padding: 2% 5% 5% 14%;
-  margin-right: 100%;
-}
-
-.submitSearch {
-  margin-left: 1%;
-}
